+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # evList plugin for glFusion - Changelog
 ## v1.4.5
 Unreleased
+- Save calendar display checkbox in session variable
 - Use the common colorpicker from LGLIB
 - Integration with the Birthdays plugin
 - Fix error when re-editing a new event due to an error